Restore scaffolding file (still needed on Darwin) and move CLI options into EntryPoint
